Authorization added

+++ Shop.Application/SubCategories/Commands/DeleteSubCategory/DeleteSubCategoryForCategoryCommandHandler.cs
@@ -3,13 +3,16 @@ using Microsoft.Extensions.Logging;
 using Shop.Domain.Entities;
 using Shop.Domain.Exceptions;
 using Shop.Domain.Repositories;
+using Shop.Infrastructure.Authorization;
+using Shop.Infrastructure.Authorization.Services;
 using System.Runtime.InteropServices;
 
 namespace Shop.Application.SubCategories.Commands.DeleteSubCategory
 {
     public class DeleteSubCategoryForCategoryCommandHandler(ILogger<DeleteSubCategoryForCategoryCommandHandler> logger,
         ICategoriesRepository categoriesRepository,
-        ISubCategoriesRepository subCategoriesRepository) : IRequestHandler<DeleteSubCategoryForCategoryCommand>
+        ISubCategoriesRepository subCategoriesRepository,
+        IShopAuthorizationService shopAuthorizationService) : IRequestHandler<DeleteSubCategoryForCategoryCommand>
     {
         public async Task Handle(DeleteSubCategoryForCategoryCommand request, CancellationToken cancellationToken)
         {
@@ -21,6 +24,9 @@ namespace Shop.Application.SubCategories.Commands.DeleteSubCategory
 
             var subcategory = category.SubCategories.FirstOrDefault(s => s.Id == request.subCategoryId);
             if (subcategory == null) throw new NotFoundException(nameof(SubCategory), request.subCategoryId.ToString());
+
+            if (!shopAuthorizationService.IsAuthorize(ResourceOperation.Delete))
+                throw new ForbidException();
 
             await subCategoriesRepository.Delete(subcategory);
         }
